pyproject.toml
[project]
name = "arduino"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncio>=4.0.0",
    "pyserial>=3.5",
    "websockets>=16.0",
]
